fix: projects를 key로 접근 못하는 버그 수정
- UUID는 String으로 타입캐스팅이 실패하기 때문에, T가 CustomStringConvertible을 채택하도록 함
